requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
black==22.6.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
cliff==3.10.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cmaes==0.8.2
cmd2==2.4.2
colorlog==6.6.0
commonmark==0.9.1
confection==0.0.3
configer==1.3.1
configparser==5.2.0
cycler==0.11.0
cymem==2.0.7
distlib==0.3.5
docker-pycreds==0.4.0
einops==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
filelock==3.7.1
fonttools==4.34.4
freetype-py==2.3.0
frozenlist==1.3.1
fsspec==2022.7.1
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.47.0
huggingface-hub==0.8.1
human-body-prior==0.8.5.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
identify==2.5.3
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
murmurhash==1.0.9
mypy-extensions==0.4.3
networkx==2.8.5
nodeenv==1.7.0
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
optuna==2.10.1
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.2
pbr==5.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
preshed==3.0.8
prettytable==3.3.0
promise==2.3
protobuf==3.19.4
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyDeprecate==0.3.2
pyglet==2.0a4
Pygments==2.12.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrender==0.1.39
pyrootutils==1.0.4
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.6.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
sentry-sdk==1.9.0
setproctitle==1.3.1
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
smplx==0.1.28
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.39
srsly==2.4.5
stevedore==4.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2022.8.8
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchgeometry==0.1.2
torchmetrics==0.7.0
torchvision==0.14.0
tqdm==4.64.0
transformers==4.21.1
transforms3d==0.3.1
trimesh==3.13.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.3
wandb==0.13.0
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.1
yarl==1.8.1
zipp==3.8.1
